Add wrapper function for m2m_aligner

+++ src/aligner4/aligner.py
@@ -40,7 +40,36 @@ class _M2MAlignerService:
     '''
     A class with utility functions to run m2m-aligner and extract its output.
     '''
-    def _align_file_content(formatted_content: str, delete=True) -> Mapping[str, str]:
+    logger = configure_logger('aligner4.aligner._M2MAlignerService')
+
+    async def m2m_aligner(**kwargs) -> None:
+        """
+        Runs the M2M aligner with the provided arguments.
+        
+        Parameters:
+            kwargs (dict): A dictionary of arguments to be passed to the M2M aligner.
+        """
+        args = []
+        for key, value in kwargs.items():
+            args.append(('-' if key in ['i', 'o'] else '--') + key)
+            if value is not True:
+                args.append(str(value))
+
+        cmd = [
+            M2M_ALIGNER_EXECUTABLE,
+            *args
+        ]
+
+        result = subprocess.run(cmd, cwd=M2M_ALIGNER_CONTAINER_DIR, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+
+        if result == 1:
+            _M2MAlignerService.logger.error(result.stderr.decode('utf-8'))
+        else:
+            _M2MAlignerService.logger.info(result.stderr.decode('utf-8'))
+
+        return result
+
+    async def _align_file_content(formatted_content: str, delete=True) -> Mapping[str, str]:
         """
         Runs the M2M aligner with the provided formatted content, and returns the aligned and unaligned data.
         
@@ -60,30 +89,35 @@ class _M2MAlignerService:
         with open(input_file_path, 'x') as file:
             file.write(formatted_content)
 
-        subprocess.run(
-            f"'{M2M_ALIGNER_EXECUTABLE}' --delX --maxX 2 --maxY 2 --init '{M2M_ALIGNER_MODEL}' --alignerIn '{M2M_ALIGNER_MODEL}' -o '{output_file_path}' -i '{input_file_path}'", cwd=M2M_ALIGNER_CONTAINER_DIR, 
-            shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL
-        )
         # - - - - - - - - - - - - - - - - - - - - - - - - #
 
-        aligned_data = ''
+        result = await _M2MAlignerService.m2m_aligner(
+            init=M2M_ALIGNER_MODEL,
+            maxX=2,
+            maxY=2,
+            alignerIn=M2M_ALIGNER_MODEL,
+            o=output_file_path,
+            i=input_file_path
+        )
+        # - - - - - - - - - - - - - - - - - - - - - - - - #
+        if result != 1:
+            aligned_data = ''
 
-        with open(output_file_path, 'r') as file:
-            aligned_data = file.read()[:-1]
+            with open(output_file_path, 'r') as file:
+                aligned_data = file.read()[:-1]
 
-        with open(unaligned_data_path, 'r') as file:
-            unaligned_data = file.read()
-        
-        if delete:
-            os.remove(input_file_path)
-            os.remove(output_file_path)
-            os.remove(unaligned_data_path)
+            with open(unaligned_data_path, 'r') as file:
+                unaligned_data = file.read()
+            
+            if delete:
+                os.remove(input_file_path)
+                os.remove(output_file_path)
+                os.remove(unaligned_data_path)
 
         return {
             "aligned_data": _M2MAlignerService._split_aligner_output(aligned_data),
             "unaligned_data": unaligned_data
         }
-    
 
     @staticmethod
     def _split_aligner_output(aligned_data: str):
